functions/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
tiktoken==0.3.3
azure.ai.formrecognizer==3.2.1
azure-storage-blob==12.16.0
azure-core == 1.26.4
beautifulsoup4 == 4.12.2
lxml == 4.9.2
mammoth == 1.5.1
azure-cosmos == 4.3.1
azure-storage-queue == 12.6.0
nltk == 3.8.1
tenacity == 8.2.2
azure-identity == 1.13.0
